+++ indexclip.js
@@ -1,8 +1,10 @@
-import { AppRegistry, SafeAreaView } from 'react-native'
-import React from 'react'
-import { ThemeProvider } from 'styled-components/native'
-import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client'
+import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client'
 import { CONFIG } from '@config/Config'
+import { NavigationContainer } from '@react-navigation/native'
+import React from 'react'
+import { AppRegistry, SafeAreaView } from 'react-native'
+import 'react-native-gesture-handler'
+import { ThemeProvider } from 'styled-components/native'
 import { AppClip } from './src/AppClip'
 import { lightTheme } from './src/styles/themes'
 
@@ -25,7 +27,9 @@ const NyxoAppClip = () => (
   <SafeAreaView>
     <ApolloProvider client={client}>
       <ThemeProvider theme={lightTheme}>
-        <AppClip />
+        <NavigationContainer>
+          <AppClip />
+        </NavigationContainer>
       </ThemeProvider>
     </ApolloProvider>
   </SafeAreaView>
