+++ ios/Podfile
@@ -15,12 +15,6 @@ target 'Nyxo' do
   use_unimodules!(exclude: ['expo-face-detector'])
 	use_flipper!
 
-
-	target 'NyxoTests' do
-    inherit! :complete
-    # Pods for testing
-  end
-
 	config = use_native_modules!
 	use_react_native!(:path => config["reactNativePath"]) 
 end 
@@ -29,6 +23,14 @@ end
 target 'Nyxo AppClip' do
 	use_react_native! 
 	pod 'RNLocalize', :path => '../node_modules/react-native-localize'
+	pod 'react-native-ultimate-config', :path => '../node_modules/react-native-ultimate-config'
+	pod 'RNScreens', :path => '../node_modules/react-native-screens'
+	pod 'RNGestureHandler', :path => '../node_modules/react-native-gesture-handler'
+	pod 'RNReanimated', :path => '../node_modules/react-native-reanimated'
+	pod 'RNCMaskedView', :path => '../node_modules/@react-native-community/masked-view'
+	pod 'react-native-safe-area-context', :path => '../node_modules/react-native-safe-area-context'
+
+
 end
