+++ package.json
@@ -133,6 +133,7 @@
     "@react-native-community/masked-view": "^0.1.6",
     "@react-native-community/netinfo": "^5.9.9",
     "@react-native-community/push-notification-ios": "^1.7.5-beta.0",
+    "@react-native-masked-view/masked-view": "^0.2.0",
     "@react-navigation/bottom-tabs": "^5.11.2",
     "@react-navigation/compat": "^5.3.10",
     "@react-navigation/core": "^5.14.4",
