+++ src/components/app-clip/Lesson.tsx
@@ -1,7 +1,8 @@
 import React, { FC } from 'react'
+import styled from 'styled-components/native'
 
-export const Lesson:FC = () => {
-	return (
-
-	)
+export const Lesson: FC = () => {
+  return <Container />
 }
+
+const Container = styled.View``
